backend/requirements.txt
Flask==3.0.3
openai==1.61.0
pymongo==4.10.1
redis==5.2.1
python-dotenv==1.0.1
requests==2.32.3
Flask-Cors==5.0.0